+++ src/ebird/codes/breeding.py
@@ -1,0 +1,42 @@
+from gettext import gettext as _
+
+BREEDING_CODES = {
+    "Fly-over": "FO",
+    "Species observed in breeding season in possible nesting habitat": "OS",
+    "Singing bird(s) present (or breeding calls heard)": "S1",
+    "Pair observed in suitable nesting habitat": "PO",
+    "Registration of territorial behaviour on at least two different days a week apart at same place": "T7",
+    "Courtship and display": "CC",
+    "Visiting probable nest-site": "VS",
+    "Agitated behaviour or anxiety calls from adults": "AB",
+    "Brood patch on adult examined in the hand": "BP",
+    "Nest-building or excavating of nest-hole": "NB",
+    "Distraction-display or injury-feigning": "DD",
+    "Used nest or eggshells found": "UN",
+    "Recently fledged young": "FL",
+    "Occupied nest": "ON",
+    "Adult carrying a faecal sac or food for young": "FY",
+    "Nests containing eggs": "NE",
+    "Nests with young seen or heard": "NY",
+}
+
+
+BREEDING_NAMES = {
+    "FO": _("Fly-over"),
+    "OS": _("Species observed in breeding season in possible nesting habitat"),
+    "S1": _("Singing bird(s) present (or breeding calls heard)"),
+    "PO": _("Pair observed in suitable nesting habitat"),
+    "T7": _("Registration of territorial behaviour on at least two different days a week apart at same place"),
+    "CC": _("Courtship and display"),
+    "VS": _("Visiting probable nest-site"),
+    "AB": _("Agitated behaviour or anxiety calls from adults"),
+    "BP": _("Brood patch on adult examined in the hand"),
+    "NB": _("Nest-building or excavating of nest-hole"),
+    "DD": _("Distraction-display or injury-feigning"),
+    "UN": _("Used nest or eggshells found"),
+    "FL": _("Recently fledged young"),
+    "ON": _("Occupied nest"),
+    "FY": _("Adult carrying a faecal sac or food for young"),
+    "NE": _("Nests containing eggs"),
+    "NY": _("Nests with young seen or heard"),
+}